hook up swap/buy/stake
