+++ packages/core-mobile/app/new/common/consts/swap.ts
