+++ packages/core-mobile/app/new/features/track/trending/components/TrendingTokensScreen.tsx
@@ -3,6 +3,13 @@ import { StyleSheet } from 'react-native'
 import { Separator } from '@avalabs/k2-alpine'
 import { CollapsibleTabs } from 'common/components/CollapsibleTabs'
 import { MarketToken } from 'store/watchlist'
+import { selectTokenVisibility } from 'store/portfolio'
+import { useSelector } from 'react-redux'
+import { selectActiveAccount } from 'store/account'
+import { selectBalanceTotalForAccount } from 'store/balance'
+import { useRouter } from 'expo-router'
+import { useNavigateToSwap } from 'features/swap/hooks/useNavigateToSwap'
+import { AVAX_TOKEN_ID } from 'common/consts/swap'
 import { TrendingTokenListItem } from './TrendingTokenListItem'
 
 const numColumns = 1
@@ -17,6 +24,34 @@ const TrendingTokensScreen = ({
   goToMarketDetail: (tokenId: string) => void
   emptyComponent: React.JSX.Element
 }): JSX.Element => {
+  const { navigate } = useRouter()
+  const { navigateToSwap } = useNavigateToSwap()
+  const activeAccount = useSelector(selectActiveAccount)
+  const tokenVisibility = useSelector(selectTokenVisibility)
+  const balanceTotal = useSelector(
+    selectBalanceTotalForAccount(activeAccount?.index ?? 0, tokenVisibility)
+  )
+  const isZeroBalance = balanceTotal === 0n
+
+  const openBuy = useCallback(() => {
+    navigate({
+      // @ts-ignore TODO: make routes typesafe
+      pathname: '/buy',
+      params: { showAvaxWarning: 'true' }
+    })
+  }, [navigate])
+
+  const onBuyPress = useCallback(
+    (initialTokenIdTo?: string) => {
+      if (isZeroBalance) {
+        openBuy()
+      } else {
+        navigateToSwap(AVAX_TOKEN_ID, initialTokenIdTo)
+      }
+    },
+    [isZeroBalance, openBuy, navigateToSwap]
+  )
+
   const renderItem = useCallback(
     ({
       item,
@@ -29,11 +64,12 @@ const TrendingTokensScreen = ({
         <TrendingTokenListItem
           token={item}
           index={index}
+          onBuyPress={onBuyPress}
           onPress={() => goToMarketDetail(item.id)}
         />
       )
     },
-    [goToMarketDetail]
+    [goToMarketDetail, onBuyPress]
   )
 
   const renderSeparator = useCallback((): JSX.Element => {
